fare_modeling/min_requirements.txt
joblib==0.14.0
pandas==0.25.2
scikit-learn==0.21.3
sklearn2pmml==0.49.1